moved execute and verify functions into the root
