+++ src/processor/mod.rs
@@ -1,8 +1,6 @@
-use log::debug;
-use std::{ cmp, time::Instant };
+use std::cmp;
 use crate::math::{ F128, FiniteField };
-use crate::stark::{ self, ProofOptions, StarkProof, MAX_OUTPUTS, MIN_TRACE_LENGTH };
-use crate::utils::{ as_bytes, Accumulator };
+use crate::stark::{ MIN_TRACE_LENGTH };
 
 mod decoder;
 use decoder::Decoder;
@@ -16,27 +14,16 @@ pub use program::{ Program, ProgramInputs };
 pub mod opcodes;
 pub mod assembly;
 
-#[cfg(test)]
-mod tests;
-
-// TODO: transforming execute() into a fully generic version results in about 10% - 15% runtime
-// penalty (mostly in running FFT). So, keeping it non-generic for now.
-
 /// Executes the specified `program` and returns the result together with program hash
 /// and STARK-based proof of execution.
 /// 
 /// * `inputs` specify the initial stack state the with inputs[0] being the top of the stack;
 /// * `num_outputs` specifies the number of elements from the top of the stack to be returned;
-pub fn execute(program: &Program, inputs: &ProgramInputs<F128>, num_outputs: usize, options: &ProofOptions) -> (Vec<F128>, [u8; 32], StarkProof<F128>)
+pub fn execute(program: &Program, inputs: &ProgramInputs<F128>) -> Vec<Vec<u128>>
 {
-    assert!(num_outputs <= MAX_OUTPUTS, 
-        "cannot produce more than {} outputs, but requested {}", MAX_OUTPUTS, num_outputs);
-
     // TODO: use the entire program to build the trace table
     let mut program = program.execution_graph().operations().to_vec();
     pad_program(&mut program);
-
-    let now = Instant::now();
 
     // TODO: clean up
     // execute the program
@@ -61,33 +48,7 @@ pub fn execute(program: &Program, inputs: &ProgramInputs<F128>, num_outputs: usi
     let mut register_traces = decoder.into_register_trace();
     register_traces.append(&mut stack.into_register_traces());
 
-    // execute the program to create an execution trace
-    
-    //let mut trace = stark::TraceTable::new(&program, inputs, options.extension_factor());
-    let mut trace = stark::TraceTable::new2(register_traces, options.extension_factor());
-    debug!("Generated execution trace of {} registers and {} steps in {} ms",
-        trace.register_count(),
-        trace.unextended_length(),
-        now.elapsed().as_millis());
-
-    // copy the user stack state the the last step to return as output
-    let last_state = trace.get_state(trace.unextended_length() - 1);
-    let outputs = last_state.get_user_stack()[0..num_outputs].to_vec();
-
-    // construct program hash
-    let mut program_hash = [0u8; 32];
-    program_hash.copy_from_slice(as_bytes(&trace.get_program_hash()));
-
-    // generate STARK proof
-    let proof = stark::prove(&mut trace, inputs.get_public_inputs(), &outputs, options);
-    return (outputs, program_hash, proof);
-}
-
-/// Verifies that if a program with the specified `program_hash` is executed with the 
-/// provided `public_inputs` and some secret inputs, the result is equal to the `outputs`.
-pub fn verify(program_hash: &[u8; 32], public_inputs: &[F128], outputs: &[F128], proof: &StarkProof<F128>) -> Result<bool, String>
-{
-    return stark::verify(program_hash, public_inputs, outputs, proof);
+    return register_traces;
 }
 
 /// Pads the program with the appropriate number of NOOPs to ensure that:
@@ -108,13 +69,4 @@ pub fn pad_program<T: FiniteField>(program: &mut Vec<T>) {
         program.len().next_power_of_two()
     };
     program.resize(cmp::max(trace_length, MIN_TRACE_LENGTH), T::from(opcodes::NOOP));
-}
-
-/// Returns a hash value of the program.
-pub fn hash_program<T: Accumulator>(program: &[T]) -> [u8; 32] {
-    assert!(program.len().is_power_of_two(), "program length must be a power of 2");
-    assert!(program.len() >= MIN_TRACE_LENGTH, "program must consist of at least {} operations", MIN_TRACE_LENGTH);
-    assert!(program[0] == T::from(opcodes::BEGIN), "program must start with BEGIN operation");
-    assert!(program[program.len() - 1] == T::from(opcodes::NOOP), "program must end with NOOP operation");
-    return T::digest(&program[..(program.len() - 1)]);
 }
